Revert "Merge pull request #20 from Apex-Finance/Testing----10/22-11/5"
This reverts commit 981c3830f2ac2eeea7522f967f0bbaf3aa72f580, reversing
changes made to 64523fb1ddf65c3a5a4e475ae41f158cc2e42cb4.
