--- Plutus/lib/models/http_exception.dart
@@ -1,11 +1,0 @@
-class HttpException implements Exception {
-  final String message;
-
-  HttpException(this.message);
-
-  @override
-  String toString() {
-    return message;
-    // return super.toString(); // Instance of HttpException
-  }
-}
