+++ Plutus/pubspec.yaml
@@ -29,9 +29,6 @@ dependencies:
   keyboard_avoider: ^0.1.2
   auto_size_text: ^2.1.0
   percent_indicator: ^2.1.8
-  http: ^0.12.0+2
-  provider: ^4.0.0
-  shared_preferences: ^0.5.3+1
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
